requirements.txt

streamlit
pandas
openpyxl
Pillow
